Feature/#66

+++ Isolate-Island/Assets/Scripts/Runtime/Stat/Stats.cs
@@ -38,7 +38,7 @@ namespace IsolateIsland.Runtime.Stat
     {
         public int MAX_HP { get; set; }
         public int HP { get; set; }
-        public int Hungry { get; set; }
+        public float Hungry { get; set; }
 
         public int ATK { get; set; }
         public int DEF { get; set; }
